refactor(ui): address PR #393 review feedback
- DagPlayhead: stop playback and hide the synced crosshair when the
  overlay is disabled or bin metadata disappears (the component stays
  mounted while rendering null, so the play interval kept ticking);
  covered by a new fake-timer test.
- DAGChart: discard stale async layout results with a cancelled flag so
  interleaved calculateLayout calls cannot overwrite a newer layout.
- DAGLegend: merge conditional classes via cn() instead of template
  string concatenation.
- DAGNodeInfoPanel: extract DataFlowMatrix and ColorDot into their own
  files (one PascalCase component per file).
- NodeFlowBar: extract SegmentValueLabel into its own file.

DagPlayhead's relative '../lib/timeline.utils' import is kept as-is:
the '@' alias maps to the app's src/ only, and relative imports are the
package-wide convention in @quent/components.
